requirements.txt
Flask==2.3.2
PyPDF2==3.0.1
python-docx==0.8.11
nltk==3.8.1
scikit-learn==1.2.2
numpy==1.24.4
Werkzeug==2.3.7
